added permissions for ownership of items

+++ backend/src/resolvers/Mutation.js
@@ -4,6 +4,7 @@ const { randomBytes } = require('crypto');
 // takes callback functions and turns them into async based functions
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const maxAge = 1000 * 60 * 60 * 24 * 365; // one year in milliseconds
 
@@ -49,9 +50,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
-    // @TODO: check if they own item, or have permissions
-    // delete item
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
+    // check if they own item, or have permissions to delete item
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+    // delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
@@ -168,6 +176,37 @@ const Mutations = {
     });
     // return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check if they are logged in
+    if (!ctx.request.userId) {
+      throw new error('You must be logged in!');
+    }
+    // query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    // check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
